level 5 unfinished
I have an issue with the countries in the 'world score' table being shown as if they are still sorted by countries. I know what the issue is but I'm unsure of how to fix it right now. I will try to come back to this if I have time.