how to fix
I don't have the time to fix this now but I believe if I use my ScoreExtractor in the right way this will work. I need to ensure that it is fed the correct data and outputs the data with the proper names so that other code will read it and I will not need to rewrite a bunch of stuff.